kv: remove cacheSnapshot
