+++ kv/union_store.go
@@ -65,15 +65,11 @@ type unionStore struct {
 
 // NewUnionStore builds a new UnionStore.
 func NewUnionStore(snapshot Snapshot) UnionStore {
-	lazy := &lazyMemBuffer{}
-	opts := make(map[Option]interface{})
-	cacheSnapshot := NewCacheSnapshot(snapshot, lazy, options(opts))
-	bufferStore := NewBufferStore(cacheSnapshot)
 	return &unionStore{
-		BufferStore:        bufferStore,
-		snapshot:           cacheSnapshot,
-		lazyConditionPairs: lazy,
-		opts:               opts,
+		BufferStore:        NewBufferStore(snapshot),
+		snapshot:           snapshot,
+		lazyConditionPairs: &lazyMemBuffer{},
+		opts:               make(map[Option]interface{}),
 	}
 }
 
@@ -124,10 +120,42 @@ func (lmb *lazyMemBuffer) Release() {
 	lmb.mb = nil
 }
 
+// Get implements the Retriever interface.
+func (us *unionStore) Get(k Key) ([]byte, error) {
+	v, err := us.MemBuffer.Get(k)
+	if IsErrNotFound(err) {
+		if _, ok := us.opts.Get(PresumeKeyNotExists); ok {
+			err = us.markLazyConditionPair(k, nil)
+			if err != nil {
+				return nil, errors.Trace(err)
+			}
+			return nil, errors.Trace(ErrNotExist)
+		}
+	}
+	if IsErrNotFound(err) {
+		v, err = us.BufferStore.r.Get(k)
+	}
+	if err != nil {
+		return v, errors.Trace(err)
+	}
+	if len(v) == 0 {
+		return nil, errors.Trace(ErrNotExist)
+	}
+	return v, nil
+}
+
 // BatchPrefetch implements the UnionStore interface.
 func (us *unionStore) BatchPrefetch(keys []Key) error {
 	_, err := us.snapshot.BatchGet(keys)
 	return errors.Trace(err)
+}
+
+// markLazyConditionPair marks a kv pair for later check.
+func (us *unionStore) markLazyConditionPair(k Key, v []byte) error {
+	if len(v) == 0 {
+		return errors.Trace(us.lazyConditionPairs.Delete(k))
+	}
+	return errors.Trace(us.lazyConditionPairs.Set(k, v))
 }
 
 // CheckLazyConditionPairs implements the UnionStore interface.
